+++ packages/eyes-universal/src/universal-server.ts
@@ -1,15 +1,22 @@
 import type * as types from '@applitools/types'
 import type {Socket} from './socket'
-import type {Driver, Context, Element, Selector} from './spec-driver'
+import type {
+  Driver as CustomDriver,
+  Context as CustomContext,
+  Element as CustomElement,
+  Selector as CustomSelector,
+} from './spec-driver/custom'
+import type {Driver as WDDriver, Element as WDElement, Selector as WDSelector} from './spec-driver/webdriver'
 import os from 'os'
 import path from 'path'
 import {makeSDK, checkSpecDriver} from '@applitools/eyes-sdk-core'
 import {makeLogger} from '@applitools/logger'
-import {makeSpec, webdriverSpec} from './spec-driver'
 import {makeHandler} from './handler'
 import {makeSocket} from './socket'
 import {makeRefer} from './refer'
 import {withTracker} from './debug/tracker'
+import {makeSpec} from './spec-driver/custom'
+import * as webdriverSpec from './spec-driver/webdriver'
 
 const IDLE_TIMEOUT = 900000 // 15min
 const LOG_DIRNAME = path.resolve(os.tmpdir(), `applitools-${new Date().toISOString()}`)
@@ -34,7 +41,8 @@ export async function makeServer({debug = false, idleTimeout = IDLE_TIMEOUT, ...
     const refer = makeRefer()
     const socket = withTracker({
       debug,
-      socket: makeSocket(client) as types.ServerSocket<Driver, Context, Element, Selector> & Socket,
+      socket: makeSocket(client) as types.ServerSocket<CustomDriver, CustomContext, CustomElement, CustomSelector> &
+        Omit<Socket, 'command' | 'request'>,
     })
 
     clearTimeout(idle)
@@ -55,30 +63,41 @@ export async function makeServer({debug = false, idleTimeout = IDLE_TIMEOUT, ...
       return {logsDir: LOG_DIRNAME}
     })
 
-    const sdk = socket.create('Core.makeSDK', ({name, version, cwd, commands, protocol}) => {
-      return makeSDK<Driver, Context, Element, Selector>({
+    const sdkPromise = socket.create('Core.makeSDK', ({name, version, cwd, commands, protocol}) => {
+      const spec = protocol === 'webdriver' ? webdriverSpec : makeSpec({socket, commands})
+      return makeSDK<
+        CustomDriver | WDDriver,
+        CustomContext | WDDriver,
+        CustomElement | WDElement,
+        CustomSelector | WDSelector
+      >({
         name: `eyes-universal/${name}`,
         version: `${require('../package.json').version}/${version}`,
         cwd,
-        spec: protocol === 'webdriver' ? webdriverSpec : makeSpec({socket, commands}),
+        spec,
         VisualGridClient: require('@applitools/visual-grid-client'),
       })
     })
 
     socket.command('Core.makeManager', async config => {
+      const sdk = await sdkPromise
       const managerRef = refer.ref(await sdk.makeManager(config))
       return managerRef
     })
     socket.command('Core.getViewportSize', async ({driver}) => {
+      const sdk = await sdkPromise
       return sdk.getViewportSize({logger, driver})
     })
     socket.command('Core.setViewportSize', async ({driver, size}) => {
+      const sdk = await sdkPromise
       return sdk.setViewportSize({logger, driver, size})
     })
     socket.command('Core.closeBatches', async ({settings}) => {
+      const sdk = await sdkPromise
       return sdk.closeBatches({logger, settings})
     })
     socket.command('Core.deleteTest', async ({settings}) => {
+      const sdk = await sdkPromise
       return sdk.deleteTest({logger, settings})
     })